On windows use upstream recipes
For compatibility with Fluendo SDK, some recipes where downgraded
since the new versions do not export the library version correctly.
They use 0.0.0 starting with the meson migration and the dynamic linker
fails sincei to resolve Fluendo SDK deps which was linked against the
older versions and it requires a version higher thatn 1700.1.3 for example.